refactor: clean up backend indexing module
