+++ backend/indexing/evm/scripts/prepare-subgraph.mjs
@@ -2,7 +2,7 @@
 /**
  * Prepare subgraph.yaml + src/addresses.ts + abis/ from Hardhat artifacts.
  *
- * Usage (from backend/indexing/evm/subgraph):
+ * Usage (from backend/indexing/evm):
  *   pnpm configure
  *
  * Env (optional — auto-read from ignition/deployments when present):
