+++ backend/indexing/evm/src/battle.ts
