+++ backend/indexing/evm/src/inventory.ts
